Silence lots of warnings in lib/DtHelp.
* 64bit warnings
* Missing prototypes
* explicit parentheses / brackets

There's still some work to be done, but, as always, it's a start ...

+++ cde/lib/DtHelp/CanvasI.h
@@ -130,7 +130,7 @@ typedef	struct	_dtCvSearchData {
 
 typedef	struct _dtCanvasStruct {
 	int		error;
-	int		txt_cnt;	/* maximum used in txt_list	*/
+	long		txt_cnt;	/* maximum used in txt_list	*/
 	int		txt_max;	/* maximum in txt_list		*/
 
 	int		line_cnt;	/* maximum used in line_lst	*/
